requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==24.1a1
click==8.1.7
Django==4.0.10
djangorestframework==3.13.1
frozenlist==1.4.1
gunicorn==20.1.0
idna==3.6
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
whitenoise==6.0.0
yarl==1.9.4
